+++ src/lib/classes/Utils/NetworkInformation/index.ts
@@ -1,4 +1,4 @@
 export { NetworkInformation } from './NetworkInformation'
 export { NetworkInformationPC } from './classes/NetworkInformationPC'
-export { NetworkInformationAbstract } from './classes/types/types.abscruct'
+// export { NetworkInformationAbstract } from './classes/types/types.abscruct'
 export { NetworkInformationCordova, type CordovaNetworkStatus } from './classes/NetworkInformationCordova'